Allow compilation with OPENSSL_NO_DEPRECATED
Updated the calls to openssl deprecated API, so that squid can be
compiled with openssl built with the OPENSSL_NO_DEPRECATED option.

Most* of the calls in the main code were updated to the openssl 1.1 API,
and the wrappers were moved into compat/openssl.h.
* The exceptions to the rule above were: (a) the BIO table setup,
which would need many calls, instead of a simple static struct;
(b) library initialization, which is optional in openssl >= 1.1, and
mandatory for < 1.1; (c) RSA key generation callback, which is absent in
openssl >= 1.1.

Applied changes from #265-Libressl support: Removed const from
*useMethod in src/ssl/bio.cc

Signed-off-by: Eneas U de Queiroz <cote2004-github@yahoo.com>